% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLDAShiny.R
\name{MLDAShiny}
\alias{MLDAShiny}
\title{Launch 'MLDAShiny' Interface}
\usage{
MLDAShiny()
}
\value{
Nothing
}
\description{
MLDAShiny() loads interactive user interface built using R 'shiny'.
}
\details{
Start MLDAShiny

The interactive user interface is to provide an easy way for linear discriminant analysis and downloading relevant plot.
}
\examples{
if(interactive()){
library(rmarkdown)
MLDAShiny()
}
}
\keyword{MLDAShiny}
